rename style.js to modalStyle.js
